% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEestimate.R
\name{se.estimate}
\alias{se.estimate}
\title{Performs k-fold cross-validation on a given data frame and records the standard errors for the k trials.}
\usage{
se.estimate(dat = NULL, pred.var = NULL, resamples = 10, test.pct = 0.1,
  test.size = NULL, method = "both", distrib = "continuous", timer = F,
  prog.bar = T)
}
\arguments{
\item{dat}{data frame}

\item{pred.var}{a variable from dat to be used as the response}
}
\value{
a list containing: (1) a vector of the k standard errors from the trials, (2) a 5-number summary of the standard errors, and
(3) if specified, a vecotr of the k build times for the trials
}
\description{
Performs k-fold cross-validation on a given data frame and records the standard errors for the k trials.
}
\examples{
se.estimate() ##defaults to use diamonds dataset from ggplot2
}
\seealso{
\code{\link{caret}} I only wrote se.estimate since caret wouldn't work on my laptop!
}
